Update to rand_keys
I made a lot of changes to the program. For one the the program now shows each keys 4th and 5th note along with its relative minor and diminished equivalent. I plan on adding another update to this program to add more customization along with more depth in what related keys are given to the user. Stay posted for future updates on the project.